KM260 - move delete cluster command to separate file so that we follow the same pattern of building commands
